Factor out EditProfileFormHandling trait
